projects/humansf/web_exp/requirements-online-dyna.txt
#env
jax==0.4.23
jaxlib==0.4.23
distrax==0.1.5
flax==0.8.2
numpy==1.26.4
# web
flask
flask-socketio
Pillow
tqdm
python-dotenv
google-cloud-storage
Flask-Cors
tqdm
ipdb
gunicorn
gevent